+++ sdk/tables/azure-data-tables/azure/data/tables/_shared/_table_client_base.py
@@ -10,8 +10,8 @@ except ImportError:
     from urlparse import urlparse  # type: ignore
 
 from azure.data.tables._shared._error import _validate_table_name
-from azure.data.tables._shared.base_client import parse_query
-from .base_client import StorageAccountHostsMixin
+from azure.data.tables._base_client import parse_query
+from .._base_client import StorageAccountHostsMixin
 
 
 class TableClientBase(StorageAccountHostsMixin):
