+++ sdk/tables/azure-data-tables/azure/data/tables/_shared/_table_service_client_base.py
@@ -9,8 +9,8 @@ try:
 except ImportError:
     from urlparse import urlparse  # type: ignore
 
-from azure.data.tables._shared.base_client import parse_query
-from .base_client import StorageAccountHostsMixin
+from azure.data.tables._base_client import parse_query
+from .._base_client import StorageAccountHostsMixin
 
 
 class TableServiceClientBase(StorageAccountHostsMixin):
