+++ sdk/tables/azure-data-tables/azure/data/tables/aio/_table_service_client_async.py
@@ -20,7 +20,7 @@ from azure.data.tables._generated.aio._azure_table_async import AzureTable
 from azure.data.tables._generated.models import TableServiceProperties, TableProperties, QueryOptions
 from azure.data.tables._models import service_stats_deserialize, service_properties_deserialize
 from azure.data.tables._shared.base_client_async import AsyncStorageAccountHostsMixin, AsyncTransportWrapper
-from azure.data.tables._shared.policies_async import ExponentialRetry
+from azure.data.tables.aio._policies_async import ExponentialRetry
 from azure.data.tables._shared.response_handlers import process_table_error
 from azure.data.tables.aio._table_client_async import TableClient
 from ._models import TablePropertiesPaged
