+++ sdk/tables/azure-data-tables/azure/data/tables/table_shared_access_signature.py
@@ -6,11 +6,11 @@
 from typing import Union
 
 from ._shared.models import AccountSasPermissions
-from azure.data.tables._models import TableServices
-from azure.data.tables._shared._common_conversion import _sign_string
-from azure.data.tables._shared._error import _validate_not_none
-from azure.data.tables._shared.constants import X_MS_VERSION
-from azure.data.tables._shared.shared_access_signature import _SharedAccessHelper,\
+from ._models import TableServices
+from ._shared._common_conversion import _sign_string
+from ._shared._error import _validate_not_none
+from ._shared.constants import X_MS_VERSION
+from .shared_access_signature import _SharedAccessHelper,\
     SharedAccessSignature, QueryStringConstants
